Pass in tool-specific file names in `getLoggingPaths()`
This patch allows passing in a list of file names that a tool needs to
the getLoggingPaths factory function. This way `getLoggingPaths` doesn't
need to know about the file paths needed by all the tools.

It also adds a test, ported from Flame.

For backwards compatibility, the tool-specific file name list is still
prepopulated for 'doctor' and 'bubbleprof'. Flame will pass in its own
custom list and at some point we could update doctor and bubbleprof to
do the same.