Allow passing a separate data file into the main template
This would be good for performance reasons, because the browser can pass
the sometimes 10s of megabytes of data as JSON instead of as JS, which
is much faster.

This would also allow us to avoid re-bundling the visualizer on every
profile later. We could instead bundle it once when we publish
Clinic.js packages. Bundling is not the biggest timesink but if we can
avoid that work it will still help a little.